refactor(TaskManager): Split interface and implementation
Separated the `TaskManager` interface from its implementation for better design and extensibility.

- Created `InMemoryTaskManager` class as the concrete implementation of `TaskManager`.
- Removed implementation details from `TaskManager`, leaving only method definitions.
- Updated references in `Main.java` to use the new `InMemoryTaskManager` class instead of the previous `TaskManager` implementation.
- Preserved existing task-related operations like add, update, and remove in the new class.
- Introduced support for dependency injection in `InMemoryTaskManager` with a `TaskRepository` parameter.
