+++ src/com/tasktracker/Main.java
@@ -5,6 +5,7 @@ import com.tasktracker.task.dto.RegularTaskCreationDTO;
 import com.tasktracker.task.dto.RegularTaskUpdateDTO;
 import com.tasktracker.task.dto.SubTaskCreationDTO;
 import com.tasktracker.task.dto.SubTaskUpdateDTO;
+import com.tasktracker.task.manager.InMemoryTaskManager;
 import com.tasktracker.task.manager.TaskManager;
 import com.tasktracker.task.model.enums.TaskStatus;
 import com.tasktracker.task.model.implementations.EpicTask;
@@ -28,7 +29,7 @@ public class Main {
   /** Basic CRUD operations: creation, reading, updating, printing. */
   private static void testBasicCrudOperations() {
     System.out.println("===== testBasicCrudOperations =====");
-    TaskManager tm = new TaskManager(new InMemoryTaskRepository());
+    TaskManager tm = new InMemoryTaskManager(new InMemoryTaskRepository());
 
     // Create two regular tasks
     RegularTask regTask1 =
@@ -79,7 +80,7 @@ public class Main {
   /** Additional epic scenarios: multiple subtasks, updates, and validations. */
   private static void testAdditionalEpicScenarios() {
     System.out.println("===== testAdditionalEpicScenarios =====");
-    TaskManager tm = new TaskManager(new InMemoryTaskRepository());
+    TaskManager tm = new InMemoryTaskManager(new InMemoryTaskRepository());
 
     EpicTask epicA =
         tm.addTask(new EpicTaskCreationDTO("Epic A Title Enough", "Epic A Description Enough"));
@@ -122,7 +123,7 @@ public class Main {
   /** Removing tasks one by one, including repeated deletions. */
   private static void testRemovingTasks() {
     System.out.println("===== testRemovingTasks =====");
-    TaskManager tm = new TaskManager(new InMemoryTaskRepository());
+    TaskManager tm = new InMemoryTaskManager(new InMemoryTaskRepository());
 
     // Create a regular task
     RegularTask regTaskA =
@@ -169,7 +170,7 @@ public class Main {
   /** Update epic and subtask statuses, ensuring epic status is recalculated properly. */
   private static void testUpdateEpicAndSubtaskStatus() {
     System.out.println("===== testUpdateEpicAndSubtaskStatus =====");
-    TaskManager tm = new TaskManager(new InMemoryTaskRepository());
+    TaskManager tm = new InMemoryTaskManager(new InMemoryTaskRepository());
 
     EpicTask epicB =
         tm.addTask(new EpicTaskCreationDTO("EpicB Title Enough", "EpicB Description Enough"));
@@ -214,7 +215,7 @@ public class Main {
   /** Removing tasks by class type. */
   private static void testRemoveTasksByType() {
     System.out.println("===== testRemoveTasksByType =====");
-    TaskManager tm = new TaskManager(new InMemoryTaskRepository());
+    TaskManager tm = new InMemoryTaskManager(new InMemoryTaskRepository());
 
     // Create multiple tasks of different types
     RegularTask reg1 =
@@ -256,7 +257,7 @@ public class Main {
    */
   private static void testBoundaryCases() {
     System.out.println("===== testBoundaryCases =====");
-    TaskManager tm = new TaskManager(new InMemoryTaskRepository());
+    TaskManager tm = new InMemoryTaskManager(new InMemoryTaskRepository());
 
     // Valid tasks
     RegularTask regValid =
